feat(front): redirect to asso in search
