+++ packages/front/src/views/association/index.ejs
